requirements.txt
bidict==0.22.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SocketIO==5.3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
python-engineio==4.4.1
python-socketio==5.8.0
Werkzeug==2.3.4
